Refactor code so to support another external reader family
