+++ nfc/external-acs/src/main/java/no/entur/android/nfc/external/acs/tag/PN532DefaultTechnology.java
@@ -3,23 +3,20 @@ package no.entur.android.nfc.external.acs.tag;
 import android.os.RemoteException;
 import android.util.Log;
 
-import com.acs.smartcard.ReaderException;
-
-import no.entur.android.nfc.util.ByteArrayHexStringConverter;
-import no.entur.android.nfc.external.acs.reader.PassthroughCommandException;
-import no.entur.android.nfc.external.acs.reader.ReaderCommandException;
-import no.entur.android.nfc.external.acs.reader.command.ACSIsoDepWrapper;
 import no.entur.android.nfc.external.service.tag.CommandTechnology;
+import no.entur.android.nfc.external.tag.AbstractReaderIsoDepWrapper;
+import no.entur.android.nfc.external.tag.DefaultTechnology;
+import no.entur.android.nfc.util.ByteArrayHexStringConverter;
 import no.entur.android.nfc.wrapper.TransceiveResult;
 
 public class PN532DefaultTechnology extends DefaultTechnology implements CommandTechnology {
 
 	protected static final String TAG = PN532DefaultTechnology.class.getName();
 
-	protected ACSIsoDepWrapper reader;
+	protected AbstractReaderIsoDepWrapper reader;
 	private boolean print;
 
-	public PN532DefaultTechnology(int tagTechnology, int slotNumber, ACSIsoDepWrapper reader, boolean print) {
+	public PN532DefaultTechnology(int tagTechnology, int slotNumber, AbstractReaderIsoDepWrapper reader, boolean print) {
 		super(tagTechnology, slotNumber);
 
 		this.reader = reader;
@@ -27,7 +24,7 @@ public class PN532DefaultTechnology extends DefaultTechnology implements Command
 	}
 
 	@Override
-	public TransceiveResult transceive(byte[] data, boolean raw) throws RemoteException {
+	public TransceiveResult transceive(byte[] data, boolean raw) {
 		try {
 			byte[] transceive;
 			if (!raw) {
@@ -44,7 +41,7 @@ public class PN532DefaultTechnology extends DefaultTechnology implements Command
 					Log.d(TAG, "Transceive raw request " + ByteArrayHexStringConverter.toHexString(data));
 				}
 
-				transceive = reader.transmitPassThrough(data);
+				transceive = reader.transceiveRaw(data);
 
 				if (print) {
 					Log.d(TAG, "Transceive raw response " + ByteArrayHexStringConverter.toHexString(transceive));
@@ -52,15 +49,7 @@ public class PN532DefaultTechnology extends DefaultTechnology implements Command
 			}
 
 			return new TransceiveResult(TransceiveResult.RESULT_SUCCESS, transceive);
-		} catch (PassthroughCommandException e) {
-			Log.d(TAG, "Problem sending command", e);
-
-			return new TransceiveResult(TransceiveResult.RESULT_FAILURE, null);
-		} catch (ReaderCommandException e) {
-			Log.d(TAG, "Problem sending command", e);
-
-			return new TransceiveResult(TransceiveResult.RESULT_FAILURE, null);
-		} catch (ReaderException e) {
+		} catch (Exception e) {
 			Log.d(TAG, "Problem sending command", e);
 
 			return new TransceiveResult(TransceiveResult.RESULT_FAILURE, null);
