+++ nfc/external-acs/src/main/java/no/entur/android/nfc/external/acs/tag/PN532NfcAAdapter.java
@@ -1,13 +1,13 @@
 package no.entur.android.nfc.external.acs.tag;
 
-import no.entur.android.nfc.external.acs.reader.command.ACSIsoDepWrapper;
+import no.entur.android.nfc.external.tag.AbstractReaderIsoDepWrapper;
 import no.entur.android.nfc.wrapper.tech.TagTechnology;
 
 public class PN532NfcAAdapter extends PN532DefaultTechnology {
 
 	protected static final String TAG = PN532NfcAAdapter.class.getName();
 
-	public PN532NfcAAdapter(int slotNumber, ACSIsoDepWrapper reader, boolean print) {
+	public PN532NfcAAdapter(int slotNumber, AbstractReaderIsoDepWrapper reader, boolean print) {
 		super(TagTechnology.NFC_A, slotNumber, reader, print);
 	}
