+++ nfc/external-acs/src/main/java/no/entur/android/nfc/external/acs/tag/TagUtility.java
@@ -11,10 +11,11 @@ import org.nfctools.mf.block.MfBlock;
 import org.nfctools.mf.ul.MfUlReaderWriter;
 import org.nfctools.spi.acs.AcrMfUlReaderWriter;
 
-import no.entur.android.nfc.util.ByteArrayHexStringConverter;
+import no.entur.android.nfc.external.ExternalNfcReaderCallback;
 import no.entur.android.nfc.external.acs.reader.command.ACRCommands;
 import no.entur.android.nfc.external.acs.reader.command.ACSIsoDepWrapper;
-import no.entur.android.nfc.external.ExternalNfcReaderCallback;
+import no.entur.android.nfc.external.tag.AbstractReaderIsoDepWrapper;
+import no.entur.android.nfc.util.ByteArrayHexStringConverter;
 
 public class TagUtility {
 
@@ -160,7 +161,7 @@ public class TagUtility {
 		return TagType.identifyTagType(historicalBytes);
 	}
 
-	public static byte[] getPcscUid(ACSIsoDepWrapper wrapper) {
+	public static byte[] getPcscUid(AbstractReaderIsoDepWrapper wrapper) {
 		try {
 			byte[] response = wrapper.transceive(GET_TAG_ID);
 			if (ACRCommands.isSuccessControl(response)) {
